Move Dockerfile to a passenger-nodejs infrastructure.
Relying on express.js on top of node.je 8.x is not sufficiently stable
for production purposes. It was not unusual to see the service crashing
on strange requests, making it available... and obvious attack vector.

The new Dockerfile uses passenger-nodejs as a basis for running the app
in production. Passenger is known for being production ready, and very
robust in practice.